Added a C program to compute heat index
This program combines data from relative humidity and maximum
temperature, per day and per grid point (in a rectangular CSV format) to
produce output in the same format with an index for each day/point
between 0 (no risk) and 4 (extreme danger).
The R code can run the same computation, except it's too slow. So it's
only used right now to generate the input files to the C code (the
sorted, matched humidity and temperature fies). Note also it currently
uses flooring instead of rounding of the humidity to index into the
lookup table.